Split traits::structural_impls in two.
